requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dj-rest-auth==7.0.1
Django==5.1.6
django-cors-headers==4.7.0
django-crontab==0.7.1
django-filter==25.1
django-rest-auth==0.9.5
djangorestframework==3.15.2
flatbuffers==25.2.10
gast==0.6.0
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.69.0
grpcio==1.70.0
h5py==3.13.0
httplib2==0.22.0
idna==3.10
joblib==1.4.2
keras==3.8.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
nltk==3.9.1
numpy==2.0.2
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
proto-plus==1.26.0
protobuf==5.29.3
psycopg==3.2.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
